fix: food -> product 클래스명 변경 적용

+++ backend/src/main/java/food/backend/search/dao/ProductDetailDao.java
@@ -12,7 +12,7 @@ public class ProductDetailDao {
 
     private final JdbcTemplate jdbcTemplate;
 
-    public ProductDetailDto getProductDetail(Integer foodId) {
+    public ProductDetailDto getProductDetail(Long foodId) {
         String sql = "SELECT * FROM products WHERE food_id = ?";
         return jdbcTemplate.queryForObject(sql, productRowMapper(), foodId);
     }
